pyproject.toml
[tool.poetry]
name        = "rag-examples"
version     = "0.1.0"
description = "Local RAG pipeline that embeds a folder of documents, stores them in Chroma, and chats via ChatGPT or webAI"
authors     = ["<you>"]

[tool.poetry.dependencies]
python                    = "^3.12"
openai                    = "^1.30"
python-dotenv             = "^1.0.1"
tiktoken = "^0.9.0"
langchain-community = "^0.3.27"
langchain-text-splitters = "^0.3.8"
chromadb = "^1.0.15"
pypdf = "^5.8.0"
nltk = "^3.9.1"
textblob = "^0.19.0"
llama-index = "^0.12.49"
langchain-openai = "^0.3.28"
sentence-transformers = "^5.0.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.25"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
